# WIP: Occurs-checker optimization