--- a/src/full/Agda/TypeChecking/MetaVars/Occurs.hs
+++ b/src/full/Agda/TypeChecking/MetaVars/Occurs.hs
@@ -522,7 +522,7 @@ instance Occurs Term where
           MetaV m' es -> do
             m' <- metaCheck m'
             ctx <- ask
-            let fallback = return . Right $ MetaV m' es
+            let fallback = return . Left $ MetaV m' es
             -- if the meta is projected don't prune it, return fallback
             t <- caseMaybe (allApplyElims es) (fallback) $ \ vs -> do
               -- don't attempt pruning if we're in a flexible context
@@ -532,19 +532,19 @@ instance Occurs Term where
                  if (killResult == PrunedEverything) then do
                    reportSDoc "tc.meta.prune" 40 $ "Pruned everything"
                    -- we don't have to do occurs-check after everything was pruned
-                   return $ Left v
+                   return $ Right v
                  else do
                    reportSDoc "tc.meta.prune" 40 $ "Didn't manage to prune everything"
                    -- some variables was pruned, but not all, still have to do occurs-check after
-                   return $ Right v
+                   return $ Left v
                else fallback
             case t of
               -- The arguments of a meta are in a flexible position
-              (Right (MetaV m' es')) -> (MetaV m' <$> do flexibly $ occurs es')
+              (Left (MetaV m' es')) -> (MetaV m' <$> do flexibly $ occurs es')
               -- Sometimes instantiated meta is actually not a meta itself
-              (Right t) -> occurs t
+              (Left t) -> occurs t
               -- the PrunedEverything case
-              (Left v) -> return v
+              (Right v) -> return v
           where
             -- a data or record type constructor propagates strong occurrences
             -- since e.g. x = List x is unsolvable