At the moment we do occurs-check twice on the arguments to the meta, once on line 525 and if that fails, we prune and check again:
https://github.com/agda/agda/blob/a0c0e586d2cbf9a2bc8ac3597fbf6733c98fc8f7/src/full/Agda/TypeChecking/MetaVars/Occurs.hs#L525-L545

This change introduces pruning eagerly and at the moment still does an occurs-check on the pruned arguments.
The tests suite passes.

Unless something has gone wrong with the tests and the reported number is off, we can get a (significant?) speedup.
std-lib test time goes from 12min22s to 10min2s. Cubical goes up from 28min16s to 28min53s.

<details> 
  <summary>std-lib test summary <b>without</b> changes, total "real time" is 12min </summary>
  
```
Total                        743,043ms            
Miscellaneous                  6,703ms            
Typing                        41,847ms (313,277ms)
Typing.CheckRHS              101,892ms            
Typing.CheckLHS               25,035ms  (80,181ms)
Typing.CheckLHS.UnifyIndices  55,145ms            
Typing.Generalize             27,722ms            
Typing.TypeSig                26,037ms            
Typing.OccursCheck            22,431ms            
Typing.With                   11,342ms            
Typing.InstanceSearch          1,821ms            
Coverage                     112,918ms            
Serialization                 87,394ms (105,123ms)
Serialization.Compress         8,493ms            
Serialization.BinaryEncode     6,498ms            
Serialization.Sort             1,925ms            
Serialization.BuildInterface     811ms            
DeadCode                      46,651ms            
Parsing                        5,343ms  (41,749ms)
Parsing.OperatorsExpr         30,028ms            
Parsing.OperatorsPattern       6,376ms            
Deserialization               29,869ms  (40,962ms)
Deserialization.Compaction    11,092ms            
Scoping                        8,282ms  (34,770ms)
Scoping.InverseScopeLookup    26,487ms            
Positivity                    26,094ms            
Termination                    2,370ms   (7,063ms)
Termination.RecCheck           3,518ms            
Termination.Compare              829ms            
Termination.Graph                344ms            
Highlighting                   4,792ms            
Injectivity                    1,412ms            
Import                         1,270ms            
ProjectionLikeness               212ms            
ModuleName                        40ms            
 918,679,447,304 bytes allocated in the heap
 103,950,975,624 bytes copied during GC
   1,545,617,576 bytes maximum residency (163 sample(s))
       4,712,144 bytes maximum slop
            2256 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     218520 colls,     0 par   148.770s  149.016s     0.0007s    0.0435s
  Gen  1       163 colls,     0 par   21.507s  21.509s     0.1320s    0.9957s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.000s elapsed)
  MUT     time  572.902s  (571.633s elapsed)
  GC      time  170.277s  (170.526s elapsed)
  EXIT    time    0.001s  (  0.001s elapsed)
  Total   time  743.180s  (742.160s elapsed)

  Alloc rate    1,603,555,129 bytes per MUT second

  Productivity  77.1% of total user, 77.0% of total elapsed


real	12m22,337s
user	12m15,298s
sys	0m8,060s
  ```
  
</details>

<details> 
  <summary>std-lib test summary <b>with</b> changes, total "real time" is 10min </summary>

```
Total                        603,257ms            
Miscellaneous                  5,676ms            
Typing                        33,625ms (251,263ms)
Typing.CheckRHS               76,516ms            
Typing.CheckLHS               19,681ms  (67,760ms)
Typing.CheckLHS.UnifyIndices  48,078ms            
Typing.Generalize             24,475ms            
Typing.TypeSig                21,345ms            
Typing.OccursCheck            18,014ms            
Typing.With                    7,753ms            
Typing.InstanceSearch          1,772ms            
Coverage                      96,083ms            
Serialization                 68,840ms  (84,022ms)
Serialization.Compress         7,412ms            
Serialization.BinaryEncode     5,536ms            
Serialization.Sort             1,511ms            
Serialization.BuildInterface     722ms            
DeadCode                      38,691ms            
Parsing                        4,678ms  (33,290ms)
Parsing.OperatorsExpr         23,759ms            
Parsing.OperatorsPattern       4,853ms            
Deserialization               23,470ms  (32,046ms)
Deserialization.Compaction     8,576ms            
Scoping                        6,878ms  (28,642ms)
Scoping.InverseScopeLookup    21,764ms            
Positivity                    22,084ms            
Termination                    1,790ms   (5,557ms)
Termination.RecCheck           2,877ms            
Termination.Compare              630ms            
Termination.Graph                259ms            
Highlighting                   3,696ms            
Injectivity                    1,138ms            
Import                           868ms            
ProjectionLikeness               170ms            
ModuleName                        24ms            
 918,834,967,056 bytes allocated in the heap
 103,758,285,736 bytes copied during GC
   1,543,961,000 bytes maximum residency (163 sample(s))
       4,714,728 bytes maximum slop
            2254 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     218555 colls,     0 par   117.242s  117.408s     0.0005s    0.0389s
  Gen  1       163 colls,     0 par   19.024s  19.026s     0.1167s    0.9332s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.000s elapsed)
  MUT     time  467.110s  (466.422s elapsed)
  GC      time  136.266s  (136.433s elapsed)
  EXIT    time    0.000s  (  0.004s elapsed)
  Total   time  603.377s  (602.860s elapsed)

  Alloc rate    1,967,063,123 bytes per MUT second

  Productivity  77.4% of total user, 77.4% of total elapsed


real	10m2,992s
user	9m57,259s
sys	0m6,250s
```
</details>

<details> 
  <summary>cubical test summary <b>without</b> changes, total "real time" is 28min16s </summary>
  
```
———— All done; warnings encountered ————————————————————————

/home/bohdan/delft/agda/agda/cubical/Cubical/Tactics/Reflection.agda:92,5-19
warning: -W[no]UserWarning
DEPRECATED: Use `withReduceDefs` instead of `dontReduceDefs`
when scope checking dontReduceDefs
make[2]: Leaving directory '/home/bohdan/delft/agda/agda/cubical'
make[1]: Leaving directory '/home/bohdan/delft/agda/agda/cubical'

real	28m16,383s
user	28m13,232s
sys	0m4,640s
```
  
</details>

<details> 
  <summary>cubical test summary <b>with</b> changes, total "real time" is 28min53s </summary>
  
```
———— All done; warnings encountered ————————————————————————

/home/bohdan/delft/agda/agda/cubical/Cubical/Tactics/Reflection.agda:92,5-19
warning: -W[no]UserWarning
DEPRECATED: Use `withReduceDefs` instead of `dontReduceDefs`
when scope checking dontReduceDefs
make[2]: Leaving directory '/home/bohdan/delft/agda/agda/cubical'
make[1]: Leaving directory '/home/bohdan/delft/agda/agda/cubical'

real	28m53,528s
user	28m50,743s
sys	0m4,376s
```

</details>


**TODO**:  
- [ ] re-run the tests to double-check
- [ ] check what happens if don't reset the state
- [ ] check what happens if we don't do occurs-check in the `PrunedEverything` case