@jespercockx 2023-10-06
I don't seem to get any speedup on my machine. Did you compile both times with `--optimize-heavily`?