@liesnikov 2023-10-06
Yeah, the `make install-bin`.

But it not making a difference makes more sense.