@liesnikov 2023-10-10
@andreasabel I was talking with @jespercockx and we believe it still makes sense to merge this just for code simplification purposes.
Could you perhaps take a look and let us know if you approve?