@liesnikov 2023-10-11
Thank you!

I titled the PR "optimization" because it seemed to run faster, but apparently that was a fluke.
I'll see what do we get with cost centre analysis.